Start to extract out a common shared core for 'Use Auto Prop'
