+++ src/EditorFeatures/CSharp/CSharpEditorResources.resx
@@ -126,7 +126,4 @@
   <data name="Properties" xml:space="preserve">
     <value>Properties</value>
   </data>
-  <data name="UseAutoProperty" xml:space="preserve">
-    <value>Use auto property</value>
-  </data>
 </root>
